[ISSUE #9509]♻️Restore Admin topic maintainability boundaries
